LG-3551, LG-2654: select option descriptions + package reorganization (#1991)
* reorganize select

* refactor to use inputOption

* more updates

* fix failing test

* fix failing tests

* update test

* comments

* changesets

* update story

* add internal tag

* fix generated story context

* robs line-height comment

* sean feedback

* use leafygreen provider for darkmode

* update readme

* update README and update changeset
