+++ .changeset/selfish-drinks-tap.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/input-option': patch
+---
+
+Exports `descriptionClassName` and adds `font-size: inherit;` to input descriptions.
