+++ packages/input-option/src/InputOption/index.ts
@@ -1,4 +1,5 @@
 export { InputOption } from './InputOption';
+export { descriptionClassName } from './InputOption.style';
 export type {
   BaseInputOptionProps,
   InputOptionProps,
