+++ packages/select/README.md
@@ -31,7 +31,9 @@ import { Option, OptionGroup, Select, Size } from '@leafygreen-ui/select';
   size={Size.Default}
   defaultValue="cat"
 >
-  <Option value="dog">Dog</Option>
+  <Option value="dog" description="Bark">
+    Dog
+  </Option>
   <Option value="cat">Cat</Option>
   <OptionGroup label="Less common">
     <Option value="hamster">Hamster</Option>
@@ -123,13 +125,14 @@ import { Option, OptionGroup, Select, Size } from '@leafygreen-ui/select';
 
 # Option
 
-| Prop        | Type                 | Description                                                                                           | Default                     |
-| ----------- | -------------------- | ----------------------------------------------------------------------------------------------------- | --------------------------- |
-| `children`  | `string`, `number`   | Content to appear inside of the component.                                                            |                             |
-| `className` | `string`             | Adds a className to the outermost element.                                                            |                             |
-| `glyph`     | `React.ReactElement` | Icon to display next to the option text.                                                              |                             |
-| `value`     | `string`             | Corresponds to the value passed into the `onChange` prop of `<Select />` when the option is selected. | text contents of `children` |
-| `disabled`  | `boolean`            | Prevents the option from being selectable.                                                            | `false`                     |
+| Prop          | Type                 | Description                                                                                           | Default                     |
+| ------------- | -------------------- | ----------------------------------------------------------------------------------------------------- | --------------------------- |
+| `children`    | `string`, `number`   | Content to appear inside of the component.                                                            |                             |
+| `className`   | `string`             | Adds a className to the outermost element.                                                            |                             |
+| `glyph`       | `React.ReactElement` | Icon to display next to the option text.                                                              |                             |
+| `value`       | `string`             | Corresponds to the value passed into the `onChange` prop of `<Select />` when the option is selected. | text contents of `children` |
+| `description` | `string`             | Optional descriptive text under the value                                                             |                             |
+| `disabled`    | `boolean`            | Prevents the option from being selectable.                                                            | `false`                     |
 
 # OptionGroup
