+++ packages/select/src/MenuButton/index.ts
@@ -1,0 +1,4 @@
+import MenuButton from './MenuButton';
+
+export default MenuButton;
+export { menuButtonTextClassName } from './MenuButton.styles';
