+++ packages/select/src/Option/index.ts
@@ -1,7 +1,2 @@
-export {
-  InternalOption,
-  type InternalProps,
-  Option,
-  type OptionElement,
-  type ReactEmpty,
-} from './Option';
+export { InternalOption, Option, type OptionElement } from './Option';
+export { type InternalProps, type ReactEmpty } from './Option.types';
