+++ packages/select/src/SelectContext.tsx
@@ -1,18 +1,14 @@
 import { createContext } from 'react';
 
-import { Theme } from '@leafygreen-ui/lib';
-
-import { Size } from './types';
+import { Size } from './Select/Select.types';
 
 interface SelectData {
-  theme: Theme;
   size: Size;
   open: boolean;
   disabled: boolean;
 }
 
 export const SelectContext = createContext<SelectData>({
-  theme: Theme.Light,
   size: Size.Default,
   open: false,
   disabled: false,
