+++ packages/select/src/index.ts
@@ -2,5 +2,10 @@ export { popoverClassName } from './ListMenu';
 export { menuButtonTextClassName } from './MenuButton';
 export { Option } from './Option';
 export { OptionGroup } from './OptionGroup';
-export { default as Select } from './Select';
-export { DropdownWidthBasis, type SelectProps, Size, State } from './types';
+export {
+  DropdownWidthBasis,
+  Select,
+  type SelectProps,
+  Size,
+  State,
+} from './Select';
