+++ packages/select/tsconfig.json
@@ -1,19 +1,26 @@
 {
   "extends": "@lg-tools/build/config/package.tsconfig.json",
-   "compilerOptions": {
+  "compilerOptions": {
     "declarationDir": "dist",
     "outDir": "dist",
     "rootDir": "src",
     "baseUrl": ".",
     "paths": {
-      "@leafygreen-ui/icon/dist/*": ["../icon/src/generated/*"],
-      "@leafygreen-ui/*": ["../*/src"]
+      "@leafygreen-ui/icon/dist/*": [
+        "../icon/src/generated/*"
+      ],
+      "@leafygreen-ui/*": [
+        "../*/src"
+      ]
     }
   },
   "include": [
     "src/**/*"
   ],
-  "exclude": ["**/*.spec.*", "**/*.story.*"],
+  "exclude": [
+    "**/*.spec.*",
+    "**/*.story.*"
+  ],
   "references": [
     {
       "path": "../button"
@@ -26,6 +33,9 @@
     },
     {
       "path": "../icon"
+    },
+    {
+      "path": "../input-option"
     },
     {
       "path": "../lib"
